fix some comments
